+++ src/solvers/linear/MA57Solver.hpp
@@ -6,16 +6,16 @@
 #include "LinearSolver.hpp"
 
 struct MA57Factorization {
-    int n;
-    int nnz;
-    std::vector<double> fact;
-    int lfact;
-    std::vector<int> ifact;
-    int lifact;
-    int lkeep;
-    std::vector<int>keep;
-    std::vector<int> iwork;
-    std::vector<int> info;
+   int n;
+   size_t nnz;
+   std::vector<double> fact;
+   int lfact;
+   std::vector<int> ifact;
+   int lifact;
+   int lkeep;
+   std::vector<int> keep;
+   std::vector<int> iwork;
+   std::vector<int> info;
 };
 
 /*! \class MA57Solver
@@ -26,27 +26,27 @@ struct MA57Factorization {
  */
 class MA57Solver : public LinearSolver {
 public:
-    MA57Solver();
-    virtual ~MA57Solver() = default;
+   MA57Solver();
+   virtual ~MA57Solver() = default;
 
-    short use_fortran;
+   short use_fortran;
 
-    void factorize(COOMatrix& matrix) override;
-    void do_symbolic_factorization(const COOMatrix& matrix) override;
-    void do_numerical_factorization(const COOMatrix& matrix) override;
-    void solve(std::vector<double>& rhs) override;
-    
-    int number_negative_eigenvalues() const override;
-    bool matrix_is_singular() const override;
-    int rank() const override;
+   void factorize(COOMatrix& matrix) override;
+   void do_symbolic_factorization(const COOMatrix& matrix) override;
+   void do_numerical_factorization(const COOMatrix& matrix) override;
+   void solve(std::vector<double>& rhs) override;
+
+   int number_negative_eigenvalues() const override;
+   bool matrix_is_singular() const override;
+   int rank() const override;
 
 private:
-    /* for ma57id_ (default values of controlling parameters) */
-    std::vector<double> cntl_;
-    std::vector<int> icntl_;
-    std::vector<double> rinfo_;
-    
-    MA57Factorization factorization_;
+   /* for ma57id_ (default values of controlling parameters) */
+   std::vector<double> cntl_;
+   std::vector<int> icntl_;
+   std::vector<double> rinfo_;
+
+   MA57Factorization factorization_;
 };
 
 #endif // MA57SOLVER_H
